backend/requirements-docker.txt
# Requirements Otimizados para Docker - Agent Solution BI
# Versão sem dependências pesadas desnecessárias (torch, transformers, etc)
# Build time: ~3-5 minutos (vs 20-30 minutos com requirements.txt completo)

# ============================================================================
# CORE - FastAPI Backend
# ============================================================================
fastapi==0.121.3
uvicorn[standard]==0.38.0
python-multipart==0.0.20
python-dotenv==1.2.1
pydantic==2.12.4
pydantic-settings==2.12.0
email-validator==2.0.0

# ============================================================================
# DATABASE - SQL Server + Parquet
# ============================================================================
sqlalchemy==2.0.44
alembic==1.17.2
pyodbc==5.3.0
aiosqlite==0.21.0
aioodbc==0.5.0

# Parquet
polars==1.34.0
polars-runtime-32==1.34.0
fastparquet==2024.11.0
pyarrow==16.1.0
pandas==2.2.2

# ============================================================================
# AUTHENTICATION & SECURITY
# ============================================================================
passlib[bcrypt]==1.7.4
bcrypt==5.0.0
python-jose[cryptography]==3.5.0
pyjwt[crypto]==2.10.1
cryptography==46.0.3

# ============================================================================
# LLM & AGENTS - Gemini/LangChain
# ============================================================================
# LangChain Core
langchain==0.2.16
langchain-core==0.2.38
langchain-community==0.2.16
langchain-text-splitters==0.2.4

# LangGraph removido (não utilizado e causando conflito)
# langgraph==0.1.4

# OpenAI (para Gemini via API compatível)
# OpenAI (para Gemini via API compatível)
openai==2.8.1
google-generativeai==0.8.3

# ============================================================================
# DATA PROCESSING
# ============================================================================
numpy==1.26.4
dask[array,dataframe]==2024.5.1
dask-expr==1.1.1

# ============================================================================
# VISUALIZATION
# ============================================================================
plotly==6.5.0
matplotlib==3.10.7
seaborn==0.13.2
kaleido==1.2.0
pillow==12.0.0

# ============================================================================
# CACHING & PERFORMANCE
# ============================================================================
slowapi==0.1.9
limits==5.6.0

# ============================================================================
# UTILITIES
# ============================================================================
requests==2.32.5
httpx==0.28.1
colorama==0.4.6
tqdm==4.67.1
psutil==7.1.3
pyngrok==7.5.0

# ============================================================================
# LOGGING & MONITORING
# ============================================================================
structlog==25.5.0
sentry-sdk==2.45.0

# ============================================================================
# TESTING (opcional, pode comentar em produção)
# ============================================================================
pytest==9.0.1
pytest-cov==7.0.0
pytest-timeout==2.4.0

# ============================================================================
# MCP Server (se usado)
# ============================================================================
mcp[cli]==1.22.0
mssql-mcp-server==0.1.2

# ============================================================================
# DEPENDÊNCIAS REMOVIDAS (não usadas no código)
# ============================================================================
# torch==2.9.1                    # ~2GB - NÃO USADO
# transformers==4.57.1            # ~500MB - NÃO USADO
# sentence-transformers==5.1.2    # ~300MB - NÃO USADO
# faiss-cpu==1.13.0               # ~100MB - NÃO USADO
# scikit-learn==1.7.2             # ~50MB - NÃO USADO (só usado por sentence-transformers)
# scipy==1.16.3                   # ~50MB - NÃO USADO (só usado por scikit-learn)
jinja2==3.1.4
